move parsing funcs to watcher package
